refactors InternalNoiseExtractors to factor in simulation method in an abstract class AgreementMethod
