+++ python/palin/internal_noise/double_pass.py
@@ -11,15 +11,17 @@ import numpy as np
 import os.path
 import warnings
 import ast
-from .internal_noise_extractor import InternalNoiseExtractor
+from .agreement_method import AgreementMethod
 
 
-class DoublePass(InternalNoiseExtractor):
+class DoublePass(AgreementMethod):
     ''' 
-    This class provides methods to estimate internal noise and criteria from reverse correlation data, provided the data has double pass trials. 
-    Internal noise and criteria are computed using the Ponsot/Neri double-pass simulation method, by first computing prob_agree and prob_first from double-pass data, 
-    and then simulating an ideal observer with a range of internal noise and criteria values to find what duplet of values generates prob_agree and prob_first that are closest to those observed in the actual data. 
+    This class implements an AgreementMethod to estimate internal noise and criteria from reverse correlation data. 
+    This requires that the data has double pass trials, from which it computes prob_agree and prob_first from double-pass data. 
     '''
+
+    def __str__(self): 
+        return 'Double-Pass method'
 
     @classmethod
     def extract_single_internal_noise(cls,data_df, trial_id, stim_id, feature_id, value_id, response_id, model_file, rebuild_model=False, internal_noise_range=np.arange(0,5,.1),criteria_range=np.arange(-5,5,1), n_repeated_trials=100, n_runs=10):
@@ -38,9 +40,6 @@ class DoublePass(InternalNoiseExtractor):
         internal_noise, criteria = cls.estimate_noise_criteria(prob_agree, prob_first, model_file, rebuild_model, internal_noise_range,criteria_range, n_repeated_trials, n_runs)
 
         return internal_noise,criteria
-
-    def __str__(self): 
-        return 'Double-Pass method'
 
     @classmethod
     def index_double_pass_trials(cls, data_df, trial_id='trial',double_pass_id='double_pass_id',value_id='value'):
@@ -99,62 +98,3 @@ class DoublePass(InternalNoiseExtractor):
         firsts = data_df[data_df[double_pass_id].notna()].groupby(trial_id).apply(lambda group: first_option(group, stim_id, response_id))
     
         return firsts.sum()/len(firsts)
-
-    @classmethod
-    def estimate_noise_criteria(cls,prob_agree, prob_first, model_file,rebuild_model=False, internal_noise_range=np.arange(0,5,.1),criteria_range=np.arange(-5,5,1), n_repeated_trials=100, n_runs=10): 
-        '''
-        Estimates internal noise and criteria given a measure of prob_agree and prob_first. 
-        Either uses a prebuilt model (a dataframe previously generated by @build_model and stored as a .csv file), or rebuild a new model. 
-        Searches through a range of possible internal noise and criteria values
-        '''    
-        # load model or rebuild
-        if os.path.isfile(model_file) & ~rebuild_model: 
-            model_df = pd.read_csv(model_file, index_col=0)
-        else:
-            model_df = cls.build_model(internal_noise_range, criteria_range, n_repeated_trials, n_runs)
-            model_df.to_csv(model_file)
-
-        # find internal_noise & criteria settings that minimizes distance to prob_agree and prob_first 
-        model_df['dist'] = model_df.apply(lambda row: (row.prob_agree-prob_agree)**2 + (row.prob_first-prob_first)**2, axis=1)
-
-        best_match = model_df[model_df.dist==model_df.dist.min()]
-
-        return best_match.internal_noise_std.iloc[0], best_match.criteria.iloc[0]
-
-    @classmethod
-    def build_model(cls,internal_noise_range=np.arange(0,5,.1),criteria_range=np.arange(-5,5,1), n_repeated_trials=100, n_runs=10): 
-        '''
-        Build a model that associates a range of internal noise and criteria values with their corresponding (simulated) prob_agree and prob_first.
-        This uses a simulated LinearObserver, and returns the model as a dataframe 
-        '''
-        print('Rebuilding double-pass model')
-
-        # deferred imports of the simulation modules, to avoid circular imports
-        from ..simulation.observers.linear_observer import LinearObserver
-        from ..simulation.trial import Int2Trial 
-        from ..simulation.experiments.double_pass_experiment import DoublePassExperiment
-        from ..simulation.trial import Int2Trial, Int1Trial 
-        from ..simulation.analysers import DouplePassStatistics
-        from ..simulation.simulation import Simulation as Sim
-
-        observer_params = {'kernel':[[1]],
-                   'internal_noise_std':internal_noise_range, 
-                  'criteria':criteria_range}
-        experiment_params = {'n_trials':[n_repeated_trials],
-                     'n_repeated':[n_repeated_trials],
-                     'trial_type': [Int2Trial],
-                     'n_features': [1],
-                     'external_noise_std': [1]}
-        analyser_params = {}
-
-        sim = Sim(DoublePassExperiment, experiment_params,
-              LinearObserver, observer_params, 
-              DoublePassStatistics, analyser_params)
-
-        sim_df = sim.run_all(n_runs=n_runs, verbose=True)
-
-        # average measures over all runs
-        sim_df.groupby(['internal_noise_std','criteria'])[DoublePassStatistics.get_metric_names()].mean()
-        return sim_df
-
-
